Refactored some Netty Stuff (#3)
