+++ java-base/src/main/java/net/cryptic_game/backend/base/netty/client/NettyClientHandler.java
@@ -1,9 +1,9 @@
-package net.cryptic_game.backend.daemon.client;
+package net.cryptic_game.backend.base.netty.client;
 
-import io.netty.channel.unix.DomainSocketAddress;
 import net.cryptic_game.backend.base.netty.EventLoopGroupHandler;
 import net.cryptic_game.backend.base.netty.NettyCodec;
 
+import java.net.SocketAddress;
 import java.util.HashSet;
 import java.util.Set;
 
@@ -17,8 +17,8 @@ public class NettyClientHandler {
         this.eventLoopGroupHandler = new EventLoopGroupHandler();
     }
 
-    public NettyClient addClient(final String name, final DomainSocketAddress socketAddress, final NettyCodec nettyCodec) {
-        final NettyClient client = new NettyClient(name, socketAddress, this.eventLoopGroupHandler, nettyCodec);
+    public NettyClient addClient(final String name, final SocketAddress address, final boolean unixSocket, final NettyCodec nettyCodec) {
+        final NettyClient client = new NettyClient(name, address, unixSocket, this.eventLoopGroupHandler, nettyCodec);
         this.clients.add(client);
         return client;
     }
