+++ java-base/src/main/java/net/cryptic_game/backend/base/netty/server/NettyServerHandler.java
@@ -1,9 +1,8 @@
-package net.cryptic_game.backend.server.server;
+package net.cryptic_game.backend.base.netty.server;
 
 import net.cryptic_game.backend.base.netty.EventLoopGroupHandler;
 import net.cryptic_game.backend.base.netty.NettyCodec;
 
-import java.net.InetSocketAddress;
 import java.net.SocketAddress;
 import java.util.HashSet;
 import java.util.Set;
@@ -18,12 +17,8 @@ public class NettyServerHandler {
         this.eventLoopGroupHandler = new EventLoopGroupHandler();
     }
 
-    public void addServer(final String name, final String host, final int port, final NettyCodec nettyCodec) {
-        this.servers.add(new NettyServer(name, new InetSocketAddress(host, port), this.eventLoopGroupHandler, nettyCodec, false));
-    }
-
-    public void addServer(final String name, SocketAddress address, final NettyCodec nettyCodec, boolean unixSocket) {
-        this.servers.add(new NettyServer(name, address, this.eventLoopGroupHandler, nettyCodec, unixSocket));
+    public void addServer(final String name, SocketAddress address, final boolean unixSocket, final NettyCodec nettyCodec) {
+        this.servers.add(new NettyServer(name, address, unixSocket, this.eventLoopGroupHandler, nettyCodec));
     }
 
     public void start() {
