+++ java-daemon/src/main/java/net/cryptic_game/backend/daemon/App.java
@@ -4,8 +4,8 @@ import com.google.gson.JsonObject;
 import io.netty.channel.unix.DomainSocketAddress;
 import net.cryptic_game.backend.base.AppBootstrap;
 import net.cryptic_game.backend.base.config.BaseConfig;
-import net.cryptic_game.backend.daemon.client.NettyClient;
-import net.cryptic_game.backend.daemon.client.NettyClientHandler;
+import net.cryptic_game.backend.base.netty.client.NettyClient;
+import net.cryptic_game.backend.base.netty.client.NettyClientHandler;
 import net.cryptic_game.backend.daemon.client.daemon.DaemonClientCodec;
 import net.cryptic_game.backend.daemon.config.DaemonConfig;
 import org.jboss.logging.Logger;
