+++ server/src/main/java/net/cryptic_game/backend/server/App.java
@@ -4,10 +4,10 @@ import io.netty.channel.unix.DomainSocketAddress;
 import net.cryptic_game.backend.base.AppBootstrap;
 import net.cryptic_game.backend.base.api.ApiHandler;
 import net.cryptic_game.backend.base.config.BaseConfig;
+import net.cryptic_game.backend.base.netty.server.NettyServerHandler;
 import net.cryptic_game.backend.server.api.ServerApiEndpointExecutor;
 import net.cryptic_game.backend.server.config.ServerConfig;
 import net.cryptic_game.backend.server.daemon.DaemonHandler;
-import net.cryptic_game.backend.server.server.NettyServerHandler;
 import net.cryptic_game.backend.server.server.daemon.DaemonCodec;
 import net.cryptic_game.backend.server.server.http.HttpCodec;
 import net.cryptic_game.backend.server.server.websocket.WebSocketCodec;
