+++ server/src/main/java/net/cryptic_game/backend/server/server/websocket/WebSocketUtils.java
@@ -1,8 +1,8 @@
 package net.cryptic_game.backend.server.server.websocket;
 
 import com.google.gson.JsonObject;
+import net.cryptic_game.backend.base.netty.server.ServerResponseType;
 import net.cryptic_game.backend.base.utils.JsonBuilder;
-import net.cryptic_game.backend.server.server.ServerResponseType;
 
 import java.util.UUID;
